Release/1690482258

+++ components/APIKeys/index.tsx
@@ -1,13 +1,15 @@
 "use client"
 
+import clsx from "clsx"
 import { useParams } from "next/navigation"
-import { useState } from "react"
+import { useMemo, useState } from "react"
 
-import Button from "@/components/Buttons"
+import Add from "@/components/icons/Add"
+import Button, { LightButtonStyles } from "@/components/Buttons"
 import CopyToClipboard from "@/components/CopyToClipboard"
 import Loading from "@/components/Loading"
+import No2FAModal from "@/components/2fa/no2FAModal"
 import Text from "@/components/Text"
-import RightArrow from "@/components/icons/RightArrow"
 import Trash from "@/components/icons/Trash"
 import Verify2FAModal from "@/components/2fa/VerifyModal"
 
@@ -15,21 +17,69 @@ import useAuthToken from "@/hooks/useAuthToken"
 import useCreateApiKey from "@/hooks/useCreateApiKey"
 import useDeleteApiKey from "@/hooks/useDeleteApiKey"
 import useGetProjectApiKeys from "@/hooks/useGetApiKeys"
+import useGetUser from "@/hooks/useGetUser"
 import { formatDate } from "@/utils/formatDate"
+import { GatewayFetchArgs } from "@/utils/gatewayFetch"
+import { ResponseError } from "@/utils/gatewayFetch"
 
 export default function APIKeys() {
-  const [showModal, setShowModal] = useState(false)
+  const [showModal, setShowModal] = useState<boolean>(false)
+  const [showNo2FAModal, setShowNo2FAModal] = useState<boolean>(false)
   const { projectId } = useParams()
   const { data, isLoading } = useGetProjectApiKeys({
     projectId,
   })
+  const { data: user } = useGetUser()
   const sessionToken = useAuthToken()
   const createMutation = useCreateApiKey(projectId)
   const deleteMutation = useDeleteApiKey(projectId)
 
+  const [pendingRequest, setPendingRequest] = useState<string>(
+    "create" || "delete"
+  )
+  const [pendingRequestParams, setPendingRequestParams] =
+    useState<GatewayFetchArgs>()
+
+  const authCodeError: boolean = useMemo(() => {
+    const error =
+      (deleteMutation.error as ResponseError) ||
+      (createMutation.error as ResponseError)
+
+    if (!error) return false
+
+    const status: number | undefined = error?.status || error?.status
+    if (status === 403) {
+      return true
+    }
+    return false
+  }, [deleteMutation.error, createMutation.error])
+
+  const statusMessage: string = useMemo(() => {
+    if (deleteMutation.isError) {
+      return "Key not Deleted, something went wrong with your request."
+    }
+    if (deleteMutation.isSuccess) {
+      return "Key Deleted!"
+    }
+    if (createMutation.isError) {
+      return "Key not Created, something went wrong with your request."
+    }
+    if (createMutation.isSuccess) {
+      return "API Key Created!"
+    }
+    return ""
+  }, [
+    deleteMutation.isError,
+    deleteMutation.isSuccess,
+    createMutation.isError,
+    createMutation.isSuccess,
+  ])
+
   const handleCreateAccessKey = () => {
+    if (!user?.is2FAEnabled) return setShowNo2FAModal(true)
     if (sessionToken) {
-      createMutation.mutate({
+      setPendingRequest("create")
+      setPendingRequestParams({
         method: "POST",
         sessionToken,
         endpointPath: `/admin/accessKey`,
@@ -38,28 +88,63 @@ export default function APIKeys() {
           roleTitle: "admin",
         }),
       })
+      setShowModal(true)
     }
   }
 
   const handleDeleteAccessKey = (keyId: string) => {
+    if (!user?.is2FAEnabled) return setShowNo2FAModal(true)
     const confirm = window.confirm("Are you sure you want to delete")
 
-    // placeholder for 2FA
-    setShowModal(true)
-
     if (confirm && sessionToken) {
-      deleteMutation.mutate({
+      setPendingRequest("delete")
+      setPendingRequestParams({
         sessionToken,
         method: "DELETE",
         endpointPath: `/admin/accessKey/${keyId}`,
       })
+      setShowModal(true)
     }
   }
+
+  const complete2FARequest = (authCode: string) => {
+    if (pendingRequest == "create") {
+      const params = {
+        ...pendingRequestParams,
+        headers: { "x-2fa-code": authCode },
+      }
+      createMutation.mutate(params as GatewayFetchArgs)
+    }
+
+    if (pendingRequest == "delete") {
+      const params = {
+        ...pendingRequestParams,
+        headers: { "x-2fa-code": authCode },
+      }
+      deleteMutation.mutate(params as GatewayFetchArgs)
+    }
+  }
+
+  const sortedData = data?.sort((a, b) => {
+    return (
+      new Date(b.AccessKey.createdAt).getTime() -
+      new Date(a.AccessKey.createdAt).getTime()
+    )
+  })
 
   return (
     <div>
       <div className="flex flex-col p-10 border-1 bg-gray-8 rounded-lg mb-10 mr-10">
-        <Text className="font-medium text-2xl pb-5">Secret keys</Text>
+        <div className="flex justify-between">
+          <Text className="font-medium text-2xl pb-5">Secret keys</Text>
+          <Button
+            className={clsx(LightButtonStyles, "flex items-center")}
+            onClick={handleCreateAccessKey}
+          >
+            <Add className="h-4 w-4 mr-4" />
+            Create a new key
+          </Button>
+        </div>
         <p className="font-small text-gray-3 text-sm pb-7">
           Secret keys are used for API endpoint authentication.
         </p>
@@ -71,7 +156,7 @@ export default function APIKeys() {
                 <Loading className=" w-1/4 h-4" />
               </div>
             ))
-          ) : data && data.length ? (
+          ) : sortedData && sortedData.length ? (
             <div className="grid-cols-3 pb-5 hidden lg:grid">
               <div className="col-span-2 grid grid-cols-2 gap-x-8">
                 <Text className="font-small text-gray-3 text-sm pb-3">Key</Text>
@@ -86,8 +171,8 @@ export default function APIKeys() {
               There are currently no API keys for this project
             </p>
           )}
-          {data &&
-            data.map(({ AccessKey }) => {
+          {sortedData &&
+            sortedData.map(({ AccessKey }) => {
               return (
                 <div
                   className="flex lg:grid grid-cols-3 pb-5"
@@ -121,18 +206,28 @@ export default function APIKeys() {
               )
             })}
         </div>
-        <div className="flex flex-row text-blue-1 items-center">
-          <Button
-            className="font-normal text-normal pr-1"
-            onClick={handleCreateAccessKey}
-          >
-            Create a new key
-          </Button>
-          <RightArrow className="w-3 h-3" />
-        </div>
       </div>
 
-      <Verify2FAModal show={showModal} closeModal={() => setShowModal(false)} />
+      <Verify2FAModal
+        show={showModal}
+        closeModal={() => setShowModal(false)}
+        onAuthCode={complete2FARequest}
+        error={
+          (deleteMutation.isError || createMutation.isError) && !authCodeError
+        }
+        authCodeError={authCodeError}
+        success={deleteMutation.isSuccess || createMutation.isSuccess}
+        loading={deleteMutation.isLoading || createMutation.isLoading}
+        reset={() => {
+          deleteMutation.reset()
+          createMutation.reset()
+        }}
+        statusMessage={statusMessage}
+      />
+      <No2FAModal
+        show={showNo2FAModal}
+        closeModal={() => setShowNo2FAModal(false)}
+      />
     </div>
   )
 }
